Added check to not delete the Desktop.ini file in shell:startup

+++ Migration_RadAdmin/Managers/ProcessManager.cs
@@ -48,8 +48,11 @@ namespace Migration_RadAdmin.Processes
 
                     foreach (string file in files)
                     {
-                        System.IO.File.Delete(file); // Delete all files in the startup folder
-                        OutputManager.Log($"Deleted file in startup folder: {file}");
+                        if (Path.GetFileName(file) != "Desktop.ini")
+                        {
+                            System.IO.File.Delete(file); // Delete all files in the startup folder
+                            OutputManager.Log($"Deleted file in startup folder: {file}");
+                        }
                     }
 
                     System.IO.File.Copy(path, finalPath, true);
